refactor: give Dome transition attempts one owner (#924)
