+++ apps/desktop/src/components/extended/metaverse/useMetaverseRoomSession.ts
@@ -8,7 +8,6 @@ import type {
   GameRoomView,
   DomePhysicsSnapshotV1,
   DomeSessionInputKindV1,
-  DomeTransitionAdmissionTicketV1,
   MetaverseAssetRef,
   MetaverseColliderV1,
   MetaverseInteractionKind,
@@ -20,13 +19,6 @@ import type {
 import type { MetaverseRoomActions } from './MetaverseRoomActions';
 import type { SessionPropView } from '../MetaverseScene';
 import { createDomeInteractionInput, persistentPropAsSharedObject } from './DomeSceneModel';
-import {
-  domeTransitionProgress,
-  transitionNeighborAtPosition,
-  transitionNeighborInZone,
-  transformAvatarBetweenDomes,
-  type DomeNeighborTransitionView,
-} from './DomeTransitionModel';
 import { useDomeTransitionNeighbors } from './useDomeTransitionNeighbors';
 import {
   DEFAULT_SHARED_OBJECT,
@@ -57,7 +49,7 @@ import {
   writeLastVisitedDome,
 } from './DomeEntryModel';
 import { loadAvatarCollider } from './AvatarColliderModel';
-import { recoverDomeTransitionCommit } from './DomeTransitionCommitRecovery';
+import { useDomeTransitionAttempt } from './useDomeTransitionAttempt';
 
 type UseMetaverseRoomSessionArgs = {
   actions: MetaverseRoomActions;
@@ -73,15 +65,6 @@ type UseMetaverseRoomSessionArgs = {
   activeChannelId?: string | null;
   configuredEntryInstanceId?: string | null;
   onError: (message: string | null) => void;
-};
-
-type TransitionAttempt = {
-  id: string;
-  sourceRoom: GameRoomView;
-  neighbor: DomeNeighborTransitionView;
-  phase: 'preparing' | 'provisional' | 'committing' | 'target_committed';
-  ticket: DomeTransitionAdmissionTicketV1 | null;
-  cancelled: boolean;
 };
 
 export type DomeRecoveryStatus = {
@@ -128,9 +111,6 @@ export function useMetaverseRoomSession({
   const [messageDraft, setMessageDraft] = useState('');
   const [sharedObject, setSharedObject] = useState<SharedRoomObjectV1>(DEFAULT_SHARED_OBJECT);
   const [sessionProps, setSessionProps] = useState<SessionPropView[]>([]);
-  const [transitionPreparingDirections, setTransitionPreparingDirections] = useState<Set<DomeDirection>>(
-    () => new Set()
-  );
   const [handoffTransform, setHandoffTransform] = useState<AvatarTransform | null>(null);
   const [lastSentSeq, setLastSentSeq] = useState(0);
   const [recoveringUntil, setRecoveringUntil] = useState(0);
@@ -147,7 +127,6 @@ export function useMetaverseRoomSession({
   const localPeerSeed = useId().replaceAll(':', '');
   const localPeerId = `${syncStatus.discovery.local_endpoint_id || syncStatus.local_author_pubkey || 'local'}:${localPeerSeed}`;
   const lastSentTransformRef = useRef<AvatarTransform | null>(null);
-  const transitionAttemptRef = useRef<TransitionAttempt | null>(null);
   const entryAttemptKeyRef = useRef<string | null>(null);
   const entryContextKeyRef = useRef<string | null>(null);
   const entryAutoDisabledRef = useRef(false);
@@ -268,18 +247,6 @@ export function useMetaverseRoomSession({
     syncStatus,
   ]);
   const knownPeerCount = Object.keys(remoteTransforms).length;
-  const transitionBoundaryStates = useMemo(() => {
-    const states: Partial<Record<DomeDirection, DomeBoundaryStateV1>> = {};
-    for (const neighbor of transitionNeighbors) {
-      states[neighbor.direction] = domeRecovery.state === 'offline'
-        ? 'offline'
-        : transitionPreparingDirections.has(neighbor.direction)
-        ? 'loading'
-        : neighbor.boundaryState;
-    }
-    return states;
-  }, [domeRecovery.state, transitionNeighbors, transitionPreparingDirections]);
-
   const nextSessionSequence = useCallback((instanceId: string, suggested = Date.now()) => {
     const next = Math.max(suggested, (sessionSequenceByInstanceRef.current.get(instanceId) ?? 0) + 1);
     sessionSequenceByInstanceRef.current.set(instanceId, next);
@@ -624,170 +591,38 @@ export function useMetaverseRoomSession({
     resetBackendEventCursor();
   }
 
-  function setTransitionPreparing(direction: DomeDirection, preparing: boolean) {
-    setTransitionPreparingDirections((current) => {
+  const applyTransitionHandoff = useCallback((
+    sourceRoom: GameRoomView, targetRoom: GameRoomView, targetTransform: AvatarTransform
+  ) => {
+    setHandoffTransform(targetTransform);
+    lastSentTransformRef.current = targetTransform;
+    setLastSentSeq(0);
+    setJoinedRoomIds((current) => {
       const next = new Set(current);
-      if (preparing) next.add(direction);
-      else next.delete(direction);
+      next.delete(sourceRoom.room_id);
+      next.add(targetRoom.room_id);
       return next;
     });
-  }
+    setSelectedRoomId(targetRoom.room_id);
+  }, []);
 
-  const abortTransitionAttempt = useCallback(async (attempt: TransitionAttempt) => {
-    if (attempt.phase === 'committing' || attempt.phase === 'target_committed') return;
-    attempt.cancelled = true;
-    if (attempt.ticket) {
-      await actions.abortTransition(attempt.ticket).catch(() => undefined);
+  const { transitionPreparingDirections, requestTransitionAbort, handleTransitionTransform } = useDomeTransitionAttempt({
+    actions, admittedRoom, localAuthorPubkey: syncStatus.local_author_pubkey, localPeerId,
+    transitionNeighbors, setTransitionNeighbors, submitInputForRoom,
+    onHandoff: applyTransitionHandoff, onError,
+  });
+
+  const transitionBoundaryStates = useMemo(() => {
+    const states: Partial<Record<DomeDirection, DomeBoundaryStateV1>> = {};
+    for (const neighbor of transitionNeighbors) {
+      states[neighbor.direction] = domeRecovery.state === 'offline'
+        ? 'offline'
+        : transitionPreparingDirections.has(neighbor.direction)
+        ? 'loading'
+        : neighbor.boundaryState;
     }
-    await submitInputForRoom(attempt.sourceRoom, {
-      type: 'abort_transition',
-      transition_id: attempt.id,
-    }).catch(() => undefined);
-    if (transitionAttemptRef.current === attempt) {
-      transitionAttemptRef.current = null;
-    }
-    setTransitionPreparingDirections((current) => {
-      const next = new Set(current);
-      next.delete(attempt.neighbor.direction);
-      return next;
-    });
-  }, [actions, submitInputForRoom]);
-
-  function beginTransitionAttempt(neighbor: DomeNeighborTransitionView) {
-    if (!admittedRoom?.metaverse || transitionAttemptRef.current) return;
-    const transitionId = `dome-transition-${globalThis.crypto?.randomUUID?.() ?? Date.now()}`;
-    const attempt: TransitionAttempt = {
-      id: transitionId,
-      sourceRoom: admittedRoom,
-      neighbor,
-      phase: 'preparing',
-      ticket: null,
-      cancelled: false,
-    };
-    transitionAttemptRef.current = attempt;
-    setTransitionPreparing(neighbor.direction, true);
-    void (async () => {
-      try {
-        await submitInputForRoom(attempt.sourceRoom, {
-          type: 'prepare_transition',
-          transition_id: attempt.id,
-          direction: neighbor.direction,
-        });
-        if (attempt.cancelled) return;
-        attempt.ticket = await actions.prepareTransition({
-          transition_id: attempt.id,
-          connection_id: neighbor.connectionId,
-          topology_digest: neighbor.topologyDigest,
-          spatial_context: attempt.sourceRoom.metaverse!.spatial_context,
-          source_instance_id: attempt.sourceRoom.metaverse!.instance_id,
-          source_instance_generation: attempt.sourceRoom.metaverse!.instance_generation,
-          target_instance_id: neighbor.room.metaverse!.instance_id,
-          target_instance_generation: neighbor.room.metaverse!.instance_generation,
-          participant_pubkey: syncStatus.local_author_pubkey,
-          direction: neighbor.direction,
-          requested_at: Date.now(),
-        });
-        if (attempt.cancelled) {
-          await abortTransitionAttempt(attempt);
-          return;
-        }
-        attempt.phase = 'provisional';
-        setTransitionPreparing(neighbor.direction, false);
-      } catch (transitionError) {
-        await abortTransitionAttempt(attempt);
-        setTransitionNeighbors((current) => current.map((candidate) =>
-          candidate.connectionId === neighbor.connectionId
-            ? { ...candidate, boundaryState: 'error' }
-            : candidate
-        ));
-        onError(
-          transitionError instanceof Error
-            ? transitionError.message
-            : 'Dome transition preparation failed'
-        );
-      }
-    })();
-  }
-
-  function commitTransitionAttempt(attempt: TransitionAttempt, transform: AvatarTransform) {
-    if (attempt.phase !== 'provisional' || attempt.cancelled || !attempt.ticket) return;
-    attempt.phase = 'committing';
-    const targetPosition = transformAvatarBetweenDomes(transform.position, attempt.neighbor.relativeCoordinateCm);
-    const targetTransform: AvatarTransform = {
-      ...transform, roomId: attempt.neighbor.room.room_id, seq: 0,
-      position: targetPosition, sentAt: Date.now(),
-    };
-    void (async () => {
-      const recovery = await recoverDomeTransitionCommit({
-        ticket: attempt.ticket!,
-        commit: () => actions.commitTransition(attempt.ticket!, targetPosition, transform.rotation),
-        getHosting: () => actions.getHosting(attempt.ticket!.request.spatial_context, attempt.ticket!.request.target_instance_id),
-        isCurrent: () => transitionAttemptRef.current === attempt && !attempt.cancelled,
-      });
-      if (recovery.status === 'cancelled') return;
-      if (recovery.status === 'rollback') {
-        attempt.phase = 'provisional';
-        await abortTransitionAttempt(attempt);
-        setTransitionNeighbors((current) => current.map((candidate) =>
-          candidate.connectionId === attempt.neighbor.connectionId
-            ? { ...candidate, boundaryState: 'error' }
-            : candidate
-        ));
-        onError(recovery.error instanceof Error
-          ? recovery.error.message
-          : 'Dome transition commit failed');
-        return;
-      }
-      attempt.phase = 'target_committed';
-      transitionAttemptRef.current = null;
-      setTransitionPreparing(attempt.neighbor.direction, false);
-      setHandoffTransform(targetTransform);
-      lastSentTransformRef.current = targetTransform;
-      setLastSentSeq(0);
-      setJoinedRoomIds((current) => {
-        const next = new Set(current);
-        next.delete(attempt.sourceRoom.room_id);
-        next.add(attempt.neighbor.room.room_id);
-        return next;
-      });
-      setSelectedRoomId(attempt.neighbor.room.room_id);
-      if (attempt.neighbor.room.metaverse) {
-        writeLastVisitedDome(
-          syncStatus.local_author_pubkey,
-          attempt.neighbor.room.metaverse.spatial_context,
-          attempt.neighbor.room.metaverse.instance_id
-        );
-      }
-      onError(null);
-      let sourceCompleted = false;
-      for (const retryDelay of [0, 250, 1_000]) {
-        if (retryDelay > 0) {
-          await new Promise((resolve) => window.setTimeout(resolve, retryDelay));
-        }
-        try {
-          await submitInputForRoom(attempt.sourceRoom, {
-            type: 'complete_transition',
-            transition_id: attempt.id,
-          });
-          sourceCompleted = true;
-          break;
-        } catch {
-          // The destination remains authoritative; retry only source cleanup.
-        }
-      }
-      if (sourceCompleted) {
-        const leftAt = Date.now();
-        await actions.publishRoomEvent(attempt.sourceRoom.room_id, localPeerId, leftAt, {
-          type: 'presence_leave',
-          room_id: attempt.sourceRoom.room_id,
-          peer_id: localPeerId,
-          left_at: leftAt,
-        }).catch(() => undefined);
-      } else {
-        onError('Destination committed; source Dome cleanup will require resynchronization');
-      }
-    })();
-  }
+    return states;
+  }, [domeRecovery.state, transitionNeighbors, transitionPreparingDirections]);
 
   const joinRoom = useCallback(async (
     roomId: string,
@@ -796,8 +631,7 @@ export function useMetaverseRoomSession({
   ): Promise<boolean> => {
     const room = rooms.find((candidate) => candidate.room_id === roomId);
     if (!room?.metaverse) return false;
-    const attempt = transitionAttemptRef.current;
-    if (attempt) void abortTransitionAttempt(attempt);
+    requestTransitionAbort();
     if (!preserveCurrent) {
       setHandoffTransform(null);
       setSelectedRoomId(roomId);
@@ -850,7 +684,7 @@ export function useMetaverseRoomSession({
       return false;
     }
   }, [
-    abortTransitionAttempt,
+    requestTransitionAbort,
     applyPhysicsSnapshot,
     localPeerId,
     onError,
@@ -1021,8 +855,7 @@ export function useMetaverseRoomSession({
       setSelectedRoomId(null);
       return;
     }
-    const attempt = transitionAttemptRef.current;
-    if (attempt) void abortTransitionAttempt(attempt);
+    requestTransitionAbort();
     const roomId = admittedRoom.room_id;
     const leftAt = Date.now();
     emit({ type: 'presence.leave', roomId, peerId: localPeerId, leftAt });
@@ -1060,26 +893,7 @@ export function useMetaverseRoomSession({
       rotation: transform.rotation,
       animation: transform.animation,
     }, transform.seq);
-    const attempt = transitionAttemptRef.current;
-    const inZone = transitionNeighborInZone(transform.position, transitionNeighbors);
-    if (!attempt && inZone) {
-      beginTransitionAttempt(inZone);
-      return;
-    }
-    if (!attempt) return;
-    if (
-      attempt.phase !== 'committing' &&
-      domeTransitionProgress(transform.position, attempt.neighbor.direction) <= 0
-    ) {
-      void abortTransitionAttempt(attempt);
-      return;
-    }
-    const crossed = transitionNeighborAtPosition(previous?.position ?? null, transform.position, [
-      attempt.neighbor,
-    ]);
-    if (crossed && attempt.phase === 'provisional') {
-      commitTransitionAttempt(attempt, transform);
-    }
+    handleTransitionTransform(transform, previous);
   }
 
   function handleSendMessage(event: FormEvent<HTMLFormElement>) {
